+++ include/faceveil/ProcessorWorker.hpp
@@ -13,6 +13,7 @@
 namespace faceveil
 {
     class ScrfdFaceDetector;
+    class PlateDetector;
 
     class ProcessorWorker final : public QObject
     {
@@ -32,11 +33,17 @@ namespace faceveil
                         bool preserveMetadata,
                         bool reviewEnabled,
                         QObject *reviewReceiver,
-                        std::shared_ptr<ScrfdFaceDetector> cachedDetector = {});
+                        std::shared_ptr<ScrfdFaceDetector> cachedDetector = {},
+                        bool detectFaces = true,
+                        bool detectPlates = false,
+                        QString plateModelPath = {},
+                        std::shared_ptr<PlateDetector> cachedPlateDetector = {});
 
         ~ProcessorWorker() override;
 
         [[nodiscard]] std::shared_ptr<ScrfdFaceDetector> takeDetector();
+
+        [[nodiscard]] std::shared_ptr<PlateDetector> takePlateDetector();
 
     public slots:
         void process();
@@ -66,7 +73,11 @@ namespace faceveil
         bool preserveMetadata_;
         bool reviewEnabled_;
         QPointer<QObject> reviewReceiver_;
+        bool detectFaces_;
+        bool detectPlates_;
+        QString plateModelPath_;
         std::atomic<bool> cancelled_{false};
         std::shared_ptr<ScrfdFaceDetector> detector_;
+        std::shared_ptr<PlateDetector> plateDetector_;
     };
 }